2010-02-25  Alan McGovern  <amcgovern@novell.com>
	* src/runtime.cpp: FE::WidthProperty takes a double but
	  active_window->GetWidth () returns an int. Cast to a double
	  so that SetValue doesn't fail when setting the width of the
	  warning messages.

svn path=/trunk/moon/; revision=152486

+++ src/runtime.cpp
@@ -948,7 +948,7 @@ Surface::ShowIncompleteSilverlightSupportMessage ()
 	storyboard->AddHandler (Timeline::CompletedEvent, HideIncompleteSilverlightSupportMessageCallback, this);
 
 	// make the message take up the full width of the window
-	message->SetValue (FrameworkElement::WidthProperty, Value (active_window->GetWidth()));
+	message->SetValue (FrameworkElement::WidthProperty, Value ((double)active_window->GetWidth()));
 }
 
 void
@@ -997,7 +997,7 @@ Surface::ShowDrmMessage ()
 	drm_message->AddHandler (UIElement::MouseLeftButtonDownEvent, HideDrmMessageCallback, this);
 
 	// make the message take up the full width of the window
-	drm_message->SetValue (FrameworkElement::WidthProperty, Value (active_window->GetWidth()));
+	drm_message->SetValue (FrameworkElement::WidthProperty, Value ((double)active_window->GetWidth()));
 }
 
 void
@@ -1178,7 +1178,7 @@ Surface::ShowFrameRateCounter ()
 
 	
 	// make the message take up the full width of the window
-	display->SetValue (FrameworkElement::WidthProperty, Value (active_window->GetWidth()));
+	display->SetValue (FrameworkElement::WidthProperty, Value ((double)active_window->GetWidth()));
 }
 
 void
